requirements_PyTorch.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.11.0=pyhd3eb1b0_1
aiohttp=3.7.3=py38h2bbff1b_1
async-timeout=3.0.1=py38haa95532_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=py_0
blas=1.0=mkl
blinker=1.4=py38haa95532_0
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2021.1.19=haa95532_0
cachetools=4.2.1=pyhd3eb1b0_0
certifi=2020.12.5=py38haa95532_0
cffi=1.14.4=py38hcd4344a_0
chardet=3.0.4=py38haa95532_1003
click=7.1.2=pyhd3eb1b0_0
colorama=0.4.4=py_0
cryptography=2.9.2=py38h7a1dbc1_0
cudatoolkit=11.0.221=h74a9793_0
cycler=0.10.0=py38_0
decorator=4.4.2=py_0
freetype=2.10.4=hd328e21_0
fsspec=0.8.5=pyhd8ed1ab_0
future=0.18.2=py38haa244fe_3
google-auth=1.24.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
grpcio=1.31.0=py38he7da953_0
h5py=2.10.0=py38h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
ignite=0.4.2=py38_0
importlib-metadata=2.0.0=py_1
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.18.1=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
jedi=0.18.0=py38haa95532_1
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py38_0
kiwisolver=1.3.0=py38hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.13.0.1=h200bbdf_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.1.0=h56a325e_1
libuv=1.40.0=he774522_0
lz4-c=1.9.3=h2bbff1b_0
markdown=3.3.3=py38haa95532_0
matplotlib=3.3.2=haa95532_0
matplotlib-base=3.3.2=py38hba9282a_0
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.2.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
multidict=4.7.6=py38he774522_1
ninja=1.10.2=py38h6d14046_0
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
oauthlib=3.1.0=py_0
olefile=0.46=py_0
openssl=1.1.1i=h2bbff1b_0
packaging=20.8=pyhd3deb0d_0
pandas=1.2.1=py38hf11a4ad_0
parso=0.8.0=py_0
pickleshare=0.7.5=py38_1000
pillow=8.1.0=py38h4fa10fc_0
pip=20.3.3=py38haa95532_0
prompt-toolkit=3.0.8=py_0
protobuf=3.13.0.1=py38ha925a31_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.7.1=py_0
pyjwt=2.0.1=py38haa95532_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38ha925a31_4
pyreadline=2.1=py38_1
pysocks=1.7.1=py38haa95532_0
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytorch=1.7.1=py3.8_cuda110_cudnn8_0
pytorch-lightning=1.1.7=pyhd8ed1ab_0
pytz=2020.5=pyhd3eb1b0_0
pywin32=227=py38he774522_1
pyyaml=5.4.1=py38h294d835_0
pyzmq=19.0.2=py38ha925a31_1
qt=5.9.7=vc14h73c81de_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7=pyhd3eb1b0_1
scikit-learn=0.23.2=py38h47e9c7a_0
scipy=1.5.2=py38h14eb087_0
setuptools=52.0.0=py38haa95532_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=py38haa95532_0
sqlite=3.33.0=h2a8f88b_0
tensorboard=2.3.0=pyh4dce500_0
tensorboard-plugin-wit=1.6.0=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu110
tornado=6.0.4=py38he774522_1
tqdm=4.56.0=pyhd8ed1ab_0
traitlets=5.0.5=py_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.3=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=py_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yarl=1.5.1=py38he774522_0
zeromq=4.3.2=ha925a31_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
